urlShortner/app/requirements.txt
Flask
Redis
cassandra-driver
waitress